+++ src/application/component/system/DrawPanelManager.java
@@ -62,10 +62,10 @@ public class DrawPanelManager {
      * @param x x座標
      * @param y y座標
      */
-    public void transfer(double x, double y) {
-        double maxX = 0, maxY = 0;  // 上限
-        double minX = GameController.getSceneWidth() - drawPane.getWidth();    // 下限
-        double minY = GameController.getSceneHeight() - drawPane.getHeight();  // 下限
+    public void transfer(int x, int y) {
+        int maxX = 0, maxY = 0;  // 上限
+        int minX = (int)(GameController.getSceneWidth() - drawPane.getWidth());    // 下限
+        int minY = (int)(GameController.getSceneHeight() - drawPane.getHeight());  // 下限
 
         // 補正
         if ( x < minX ) { x = minX; } else if ( x > maxX ) { x = maxX; }
